deployment: make sure that we have a unique salt
This fixes a bug - every call to getPredictedSafeAddress was returning a different address due to a getRandomBytes call. Right now we can specify randombytes in the .env files and this way they the generated safe address would always be the same for the whole deployment.
In addition to that, we also now at the start of a deployment check if the safe is already deployed and if it is we abort.